telemetry: use thread pool for telemetry updates
Instead of calling the callback directly on the receiver thread we are
now using the thread pool which means that we don't block the receiver
thread if the user does something involved in the callback.

Also, by making a copy before passing the callback to the thread pool,
we enable the subscription to cancel or change itself. For instance
using `nullptr` a subscription can cancel itself again if it no longer
needs updates.